TrialOutputDemo
This update works on recording data for each pass functionality.
The moment when the asteroid either is hit by the projectile or is outside of the camera's bounds, we load whatever data we could find from the last pass into the data models and outputs a line of results to the console.

Lotsa stuff, I commented the Asteroid and TrialController classes